+++ client/src/Sass/components/_Timetable.scss
@@ -14,13 +14,13 @@
   }
 }
 
-.timesheets {
+.timetables {
   display: flex;
   flex-direction: column;
   gap: 3rem;
 }
 
-.timesheet {
+.timetable {
   //TODO: Remove
   width:61.2rem;
