+++ client/src/Sass/main.scss
@@ -14,7 +14,7 @@
 @import "components/Nav";
 @import "components/RedirectPrompt";
 @import "components/SubNav";
-@import "components/Timesheet";
+@import "components/Timetable";
 @import "components/VirtualFairSchedule";
 
 @import "components/button";
