+++ client/src/Views/MyEventsPage/MyEventsPage.tsx
@@ -12,9 +12,9 @@ import RedirectPrompt from 'Components/RedirectPrompt';
 import SubNavLink from 'Components/SubNavLink/SubNavLink';
 import SubNav from 'Components/SubNav/SubNav';
 
-import VolunteerTimesheet from './VolunteerTimesheet/VolunteerTimesheet';
-import StudentTimesheet from './StudentTimesheet/StudentTimesheet';
-import RepresentativeFairTimesheetVF from './RepresentativeFairTimesheet/RepresentativeFairTimesheetVF';
+import VolunteerTimetable from './VolunteerTimetable/VolunteerTimetable';
+import StudentTimetable from './StudentTimetable/StudentTimetable';
+import RepresentativeFairTimetableVF from './RepresentativeFairTimetable/RepresentativeFairTimetableVF';
 
 
 interface OwnProps {
@@ -45,7 +45,7 @@ class MyEventsPage extends React.Component<Props, OwnState> {
     const path: string = this.props.parentPath+subPath;
     return (
       <Route exact path={path} key={path}>
-        <VolunteerTimesheet eventTitle={eventTitle} />
+        <VolunteerTimetable eventTitle={eventTitle} />
       </Route>
     );
   }
@@ -54,7 +54,7 @@ class MyEventsPage extends React.Component<Props, OwnState> {
     const path: string = this.props.parentPath+subPath;
     return (
       <Route exact path={path} key={path}>
-        <RepresentativeFairTimesheetVF eventTitle={eventTitle} />
+        <RepresentativeFairTimetableVF eventTitle={eventTitle} />
       </Route>
     );
   }
@@ -63,7 +63,7 @@ class MyEventsPage extends React.Component<Props, OwnState> {
     const path: string = this.props.parentPath+subPath;
     return (
       <Route exact path={path} key={path}>
-        <StudentTimesheet eventTitle={eventTitle} />
+        <StudentTimetable eventTitle={eventTitle} />
       </Route>
     );
   }
@@ -77,13 +77,13 @@ class MyEventsPage extends React.Component<Props, OwnState> {
 
   private _renderVolunteerRoutes(): JSX.Element[] {
     return ([
-      this._renderVolunteerRoute(`/timesheet/portfolio-review-1`, 'Portfolio Review 1'),
-      this._renderVolunteerRoute(`/timesheet/mock-interview-1`, 'Mock Interview 1'),
-      this._renderVolunteerRoute(`/timesheet/mock-interview-2`, 'Mock Interview 2'),
-      this._renderVolunteerRoute(`/timesheet/portfolio-review-2`, 'Portfolio Review 2'),
+      this._renderVolunteerRoute(`/Timetable/portfolio-review-1`, 'Portfolio Review 1'),
+      this._renderVolunteerRoute(`/Timetable/mock-interview-1`, 'Mock Interview 1'),
+      this._renderVolunteerRoute(`/Timetable/mock-interview-2`, 'Mock Interview 2'),
+      this._renderVolunteerRoute(`/Timetable/portfolio-review-2`, 'Portfolio Review 2'),
       ...(
         this.props.user?.isRepresentative
-        ? [this._renderRepresentativeFairRoute(`/timesheet/virtual-fair`, 'Virtual Fair')]
+        ? [this._renderRepresentativeFairRoute(`/Timetable/virtual-fair`, 'Virtual Fair')]
         : []
       )
     ]);
@@ -92,13 +92,13 @@ class MyEventsPage extends React.Component<Props, OwnState> {
   private _renderVolunteerLinks(): JSX.Element {
     return (
       <>
-        {this._renderLink(`/timesheet/portfolio-review-1`, 'Portfolio Review 1')}
-        {this._renderLink(`/timesheet/mock-interview-1`, 'Mock Interview 1')}
-        {this._renderLink(`/timesheet/mock-interview-2`, 'Mock Interview 2')}
-        {this._renderLink(`/timesheet/portfolio-review-2`, 'Portfolio Review 2')}
+        {this._renderLink(`/Timetable/portfolio-review-1`, 'Portfolio Review 1')}
+        {this._renderLink(`/Timetable/mock-interview-1`, 'Mock Interview 1')}
+        {this._renderLink(`/Timetable/mock-interview-2`, 'Mock Interview 2')}
+        {this._renderLink(`/Timetable/portfolio-review-2`, 'Portfolio Review 2')}
         {
           this.props.user?.isRepresentative &&
-          this._renderLink(`/timesheet/virtual-fair`, 'Virtual Fair')
+          this._renderLink(`/Timetable/virtual-fair`, 'Virtual Fair')
         }
       </>
     );
@@ -106,20 +106,20 @@ class MyEventsPage extends React.Component<Props, OwnState> {
 
   private _renderStudentRoutes(): JSX.Element[] {
     return ([
-      this._renderStudentRoute(`/timesheet/portfolio-review-1`, 'Portfolio Review 1'),
-      this._renderStudentRoute(`/timesheet/mock-interview-1`, 'Mock Interview 1'),
-      this._renderStudentRoute(`/timesheet/mock-interview-2`, 'Mock Interview 2'),
-      this._renderStudentRoute(`/timesheet/portfolio-review-2`, 'Portfolio Review 2'),
+      this._renderStudentRoute(`/Timetable/portfolio-review-1`, 'Portfolio Review 1'),
+      this._renderStudentRoute(`/Timetable/mock-interview-1`, 'Mock Interview 1'),
+      this._renderStudentRoute(`/Timetable/mock-interview-2`, 'Mock Interview 2'),
+      this._renderStudentRoute(`/Timetable/portfolio-review-2`, 'Portfolio Review 2'),
     ]);
   }
 
   private _renderStudentLinks(): JSX.Element {
     return (
       <>
-      {this._renderLink(`/timesheet/portfolio-review-1`, 'Portfolio Review 1')}
-      {this._renderLink(`/timesheet/mock-interview-1`, 'Mock Interview 1')}
-      {this._renderLink(`/timesheet/mock-interview-2`, 'Mock Interview 2')}
-      {this._renderLink(`/timesheet/portfolio-review-2`, 'Portfolio Review 2')}
+      {this._renderLink(`/Timetable/portfolio-review-1`, 'Portfolio Review 1')}
+      {this._renderLink(`/Timetable/mock-interview-1`, 'Mock Interview 1')}
+      {this._renderLink(`/Timetable/mock-interview-2`, 'Mock Interview 2')}
+      {this._renderLink(`/Timetable/portfolio-review-2`, 'Portfolio Review 2')}
       </>
     );
   }
