+++ client/src/Views/MyEventsPage/StudentTimetable/StudentTimetableRow/StudentTimetableRow.tsx
@@ -7,7 +7,7 @@ import { msToTimeString } from 'Shared/utils';
 import Meeting from 'Shared/entityClasses/Meeting';
 import User from 'Shared/entityClasses/User';
 
-import TimesheetRowButton from "Components/TimesheetRowButton/TimesheetRowButton";
+import TimetableRowButton from "Components/TimetableRowButton/TimetableRowButton";
 
 
 interface OwnProps {
@@ -31,7 +31,7 @@ const connector = connect(mapStateToProps, mapDispatchToProps);
 
 type Props = ConnectedProps<typeof connector> & OwnProps;
 
-class StudentTimesheetRow extends React.Component<Props, OwnState> {
+class StudentTimetableRow extends React.Component<Props, OwnState> {
   public componentDidMount(): void {
 
   }
@@ -53,14 +53,14 @@ class StudentTimesheetRow extends React.Component<Props, OwnState> {
     } = this.props;
 
     return (
-      <div className="student-timesheet-row table__row table__row--student">
+      <div className="student-timetable-row table__row table__row--student">
         <div className="table__cell table__cell--time">
-          <TimesheetRowButton
+          <TimetableRowButton
             disabled={true}
             modifier={this._generateButtonColor()}
           >
             {this._generateTimeString()}
-          </TimesheetRowButton>
+          </TimetableRowButton>
         </div>
         <div className="table__cell table__cell--name">{assignee && `${assignee.firstname} ${assignee.lastname}`}</div>
         <div className="table__cell table__cell--email">{assignee && <a href={`mailto:${assignee.email}`}>{assignee.email}</a>}</div>
@@ -69,4 +69,4 @@ class StudentTimesheetRow extends React.Component<Props, OwnState> {
   }
 }
 
-export default connector(StudentTimesheetRow);
+export default connector(StudentTimetableRow);
