+++ client/src/Views/MyEventsPage/VolunteerTimetable/VolunteerTimetable.tsx
@@ -9,7 +9,7 @@ import Event from 'Shared/entityClasses/Event';
 import Meeting from 'Shared/entityClasses/Meeting';
 import User from 'Shared/entityClasses/User';
 
-import StudentTimesheetRow from './StudentTimesheetRow/StudentTimesheetRow';
+import VolunteerTimetableRow from './VolunteerTimetableRow/VolunteerTimetableRow';
 
 
 interface OwnProps {
@@ -17,6 +17,8 @@ interface OwnProps {
 }
 
 interface OwnState {
+  dispatchQueue: any;
+  isLoading: boolean;
 }
 
 type TimeOption = {
@@ -42,12 +44,12 @@ const mapStateToProps = (state: IRootState, ownProps: any) => {
 
   return {
     event,
-    isAttendingEvent,
     registrationIsOpen: event?.registrationIsOpen,
     isPreRegistration: event?.isPreRegistration,
     isPostRegistration: event?.isPostRegistration,
+    isAttendingEvent,
     users: state.userData.users,
-    meetings: state.auth.user?.findInvitedMeetings(event?.findMeetings(state.meetingData.meetings) ?? []) ?? [],
+    meetings: state.auth.user?.findOwnedMeetings(event?.findMeetings(state.meetingData.meetings) ?? []) ?? [],
 
     eventsAreStale,
     isLoading_fetchEvents,
@@ -79,7 +81,9 @@ const connector = connect(mapStateToProps, mapDispatchToProps);
 
 type Props = ConnectedProps<typeof connector> & OwnProps;
 
-class VolunteerTimesheetPR2 extends React.Component<Props, OwnState> {
+class VolunteerTimetable extends React.Component<Props, OwnState> {
+  state = {dispatchQueue: {}, isLoading: false};
+
   public componentDidMount(): void {
     if (this.props.eventsAreStale && !this.props.isLoading_fetchEvents) {
       this.props.fetchEvents();
@@ -110,19 +114,35 @@ class VolunteerTimesheetPR2 extends React.Component<Props, OwnState> {
     }
   }
 
-  private _renderTimeSlots(timeSlots: any[]): JSX.Element[] {
+  private _onSaveChanges = (): void => {
+    this.setState({ isLoading: true });
+    console.log(Object.values(this.state.dispatchQueue));
+
+    Promise.allSettled(Object.values(this.state.dispatchQueue).map((func: any) => func()))
+    .then(() => {
+      this.setState({dispatchQueue: {}, isLoading: false});
+    })
+  }
+
+  private _setReaction = (key: string, reaction: any) => {
+    this.setState({ dispatchQueue: {...this.state.dispatchQueue, [key]: reaction} });
+  }
+
+  private _renderTimeOptions(timeSlots: any[]): JSX.Element[] {
     const { meetings, event } = this.props;
     return timeSlots.map(({start_time, end_time}: TimeOption) => {
       const meeting: Meeting | null = Meeting.findByTime(meetings, start_time, end_time);
-      const assignee: User | undefined = meeting?.findOwner(this.props.users) ?? undefined;
+      const assignee: User | undefined = meeting?.findInvitee(this.props.users) ?? undefined;
       return (
         <React.Fragment key={start_time}>
-          <StudentTimesheetRow
+          <VolunteerTimetableRow
             start_time={start_time}
             end_time={end_time}
             event_id={event?.id}
             meeting={meeting}
             assignee={assignee}
+            setReaction={this._setReaction}
+            registrationIsOpen={this.props.registrationIsOpen}
           />
         </React.Fragment>
       );
@@ -141,19 +161,26 @@ class VolunteerTimesheetPR2 extends React.Component<Props, OwnState> {
     if (this.props.errors.length > 0) {
       this.props.errors.forEach((error: string) => console.error(error));
       return (
-        <div className="error">{`Failed to load${event?.title ? ` ${event.title}` : ''} timesheet`}</div>
+        <div className="error">{`Failed to load${event?.title ? ` ${event.title}` : ''} Timetable`}</div>
       );
     }
 
     if (this.props.isLoading) {
       return (
-        <div>{`Loading Student Timesheet${event?.title ? ` for ${event.title}` : ''}...`}</div>
+        <div>{`Loading Volunteer Timetable${event?.title ? ` for ${event.title}` : ''}...`}</div>
+      );
+    }
+
+    if (this.state.isLoading) {
+      return (
+        <div>Saving changes...</div>
       );
     }
 
     return (
-      <div className="student-timesheet timesheet timesheet--student">
-        <h2 className="heading-secondary">{`Student ${event?.title} Timesheet`}</h2>
+      <div className="volunteer-timetable timetable timetable--volunteer">
+        <h2 className="heading-secondary">{`Volunteer ${event?.title} Timetable`}</h2>
+
         {
           registrationIsOpen &&
             <button
@@ -166,16 +193,16 @@ class VolunteerTimesheetPR2 extends React.Component<Props, OwnState> {
         }
         {
           isPreRegistration &&
-          <p>Registration is currently not yet open. No registration changes can be made at this time.</p>
+          <p>Registration is currently not yet open. No registration changes or timeslot modifications can be made at this time.</p>
         }
         {
           isPostRegistration &&
-          <p>Registration is currently closed. No registration changes can be made at this time.</p>
+          <p>Registration is currently closed. No registration changes or timeslot modifications can be made at this time.</p>
         }
 
         {
           !isAttendingEvent && isPostRegistration &&
-          <div>{`Timesheet not available. ${registrationIsOpen ? 'Please register for this event!' : 'You did not register for this event!'}`}</div>
+          <div>{`Timetable not available. ${registrationIsOpen ? 'Please register for this event!' : 'You did not register for this event!'}`}</div>
         }
 
         {
@@ -184,16 +211,21 @@ class VolunteerTimesheetPR2 extends React.Component<Props, OwnState> {
             <div className="table">
               <div className="table__rows">
 
-                <div className="table__row table__row--student table__row--header">
+                <div className="table__row table__row--volunteer table__row--header">
                   <div className="table__cell table__cell--header">Time</div>
                   <div className="table__cell table__cell--header">Name</div>
-                  <div className="table__cell table__cell--header">Email</div>
+                  <div className="table__cell table__cell--header">Portfolio</div>
+                  <div className="table__cell table__cell--header">Resume</div>
                 </div>
 
-                {event && this._renderTimeSlots(event.createTimeSlots(20, 5))}
+                {event && this._renderTimeOptions(event.createTimeSlots(20, 5))}
 
               </div>
             </div>
+            {
+              registrationIsOpen && 
+              <button onClick={() => this._onSaveChanges()}>Save Changes</button>
+            }
           </>
         }
       </div>
@@ -201,4 +233,4 @@ class VolunteerTimesheetPR2 extends React.Component<Props, OwnState> {
   }
 }
 
-export default connector(VolunteerTimesheetPR2);
+export default connector(VolunteerTimetable);
